# lib/api-zod/src/generated/types/createSuggestionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Writers Room API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSuggestionBody {
  originalText: string;
  suggestedText: string;
  /** @nullable */
  comment?: string | null;
  submitterId: number;
}