# lib/api-zod/src/generated/types/inviteCollaboratorBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Writers Room API
 * OpenAPI spec version: 0.1.0
 */

export interface InviteCollaboratorBody {
  email: string;
  ownerId: number;
}