# lib/api-zod/src/generated/types/projectDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Writers Room API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectDetailRole } from "./projectDetailRole";
import type { ProjectDetailType } from "./projectDetailType";

export interface ProjectDetail {
  id: number;
  title: string;
  type: ProjectDetailType;
  content: string;
  ownerId: number;
  ownerName: string;
  role: ProjectDetailRole;
  createdAt: Date;
  updatedAt: Date;
}