# lib/api-zod/src/generated/types/suggestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Writers Room API
 * OpenAPI spec version: 0.1.0
 */
import type { SuggestionStatus } from "./suggestionStatus";

export interface Suggestion {
  id: number;
  projectId: number;
  submitterId: number;
  submitterName: string;
  originalText: string;
  suggestedText: string;
  /** @nullable */
  comment?: string | null;
  status: SuggestionStatus;
  /** @nullable */
  ownerNote?: string | null;
  createdAt: Date;
  updatedAt: Date;
}