# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Writers Room API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProjectBody {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  content?: string | null;
  userId: number;
}